XmlLoggingConfiguration - AutoReload available by default as before (#5782)
