+++ src/NLog.sln
@@ -65,10 +65,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NLog.RegEx", "NLog.RegEx\NL
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NLog.RegEx.Tests", "..\tests\NLog.RegEx.Tests\NLog.RegEx.Tests.csproj", "{7C77DA2A-DF9A-4F4C-91BF-A593B31D7619}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NLog.AutoReloadConfig", "NLog.AutoReloadConfig\NLog.AutoReloadConfig.csproj", "{FB2DFD9F-5EB1-4BBF-99F3-38C2E997206E}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NLog.AutoReloadConfig.Tests", "..\tests\NLog.AutoReloadConfig.Tests\NLog.AutoReloadConfig.Tests.csproj", "{47F3A0BB-E7FB-41B5-8172-B8F2DC5C2E7A}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NLog.Targets.ConcurrentFile", "NLog.Targets.ConcurrentFile\NLog.Targets.ConcurrentFile.csproj", "{C54C900A-2FAC-4482-BFFB-D4A9F6AB3619}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NLog.Targets.ConcurrentFile.Tests", "..\tests\NLog.Targets.ConcurrentFile.Tests\NLog.Targets.ConcurrentFile.Tests.csproj", "{A869B720-AB81-4AA9-94B4-12C5CF490FFE}"
@@ -167,14 +163,6 @@ Global
 		{7C77DA2A-DF9A-4F4C-91BF-A593B31D7619}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{7C77DA2A-DF9A-4F4C-91BF-A593B31D7619}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{7C77DA2A-DF9A-4F4C-91BF-A593B31D7619}.Release|Any CPU.Build.0 = Release|Any CPU
-		{FB2DFD9F-5EB1-4BBF-99F3-38C2E997206E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{FB2DFD9F-5EB1-4BBF-99F3-38C2E997206E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{FB2DFD9F-5EB1-4BBF-99F3-38C2E997206E}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{FB2DFD9F-5EB1-4BBF-99F3-38C2E997206E}.Release|Any CPU.Build.0 = Release|Any CPU
-		{47F3A0BB-E7FB-41B5-8172-B8F2DC5C2E7A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{47F3A0BB-E7FB-41B5-8172-B8F2DC5C2E7A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{47F3A0BB-E7FB-41B5-8172-B8F2DC5C2E7A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{47F3A0BB-E7FB-41B5-8172-B8F2DC5C2E7A}.Release|Any CPU.Build.0 = Release|Any CPU
 		{C54C900A-2FAC-4482-BFFB-D4A9F6AB3619}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{C54C900A-2FAC-4482-BFFB-D4A9F6AB3619}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C54C900A-2FAC-4482-BFFB-D4A9F6AB3619}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -222,8 +210,6 @@ Global
 		{38828090-4953-4EF5-929C-8063FD4BCCC9} = {194AB4BD-C817-4C59-A86C-49D89B8C3A64}
 		{A87BFA04-DBFD-44F4-8223-A02AC0E85C7F} = {194AB4BD-C817-4C59-A86C-49D89B8C3A64}
 		{7C77DA2A-DF9A-4F4C-91BF-A593B31D7619} = {194AB4BD-C817-4C59-A86C-49D89B8C3A64}
-		{FB2DFD9F-5EB1-4BBF-99F3-38C2E997206E} = {194AB4BD-C817-4C59-A86C-49D89B8C3A64}
-		{47F3A0BB-E7FB-41B5-8172-B8F2DC5C2E7A} = {194AB4BD-C817-4C59-A86C-49D89B8C3A64}
 		{C54C900A-2FAC-4482-BFFB-D4A9F6AB3619} = {194AB4BD-C817-4C59-A86C-49D89B8C3A64}
 		{A869B720-AB81-4AA9-94B4-12C5CF490FFE} = {194AB4BD-C817-4C59-A86C-49D89B8C3A64}
 		{68CE5BB3-DF9D-4774-9B66-6F7378BC04F5} = {194AB4BD-C817-4C59-A86C-49D89B8C3A64}
