+++ src/NLog/Internal/MultiFileWatcher.cs
